chore(eslint): move rule into file
